% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{invoke}
\alias{invoke}
\title{Add a method invocation to a widget object}
\usage{
invoke(widget, method, ...)
}
\arguments{
\item{widget}{an HTMLWidget}

\item{method}{Method name}

\item{...}{Arguments passed to the method in Javascript}
}
\value{
\code{widget} with the method invocations appended
}
\description{
Add a method invocation to a widget object
}
