% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Create a Maplibre-GL map widget}
\usage{
map(
  style = openstreetmap(),
  width = NULL,
  height = NULL,
  element_id = NULL,
  ...
)
}
\arguments{
\item{style}{A style object or URL to style json}

\item{width}{Width}

\item{height}{Height}

\item{element_id}{HTML Element ID}

\item{...}{Additional options passed to maplibregl constructor}
}
\description{
Create a Maplibre-GL map widget
}
