% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer.R
\name{map_layer}
\alias{map_layer}
\title{Add a layer to a map.}
\usage{
map_layer(
  widget,
  id,
  source = NULL,
  source_layer = NULL,
  type = layer_types(),
  paint = list(),
  popup_text = character(0)
)
}
\arguments{
\item{widget}{Map widget}

\item{id}{ID of layer}

\item{popup_text}{a \code{list} with at least the element \verb{$text}.
\verb{$text} must be a string which may contain
string interpolation variables (\verb{$\{...\}}).
For example, if \code{source} has a property \code{mag},
the popup text can be specified as:
\code{"Magnitude: ${mag}"}
to display the \code{mag} property.}
}
\description{
Add a layer to a map.
}
