% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint_background}
\alias{paint_background}
\alias{paint_fill}
\alias{paint_line}
\alias{paint_symbol}
\alias{paint_raster}
\alias{paint_circle}
\alias{paint_fill_extrusion}
\alias{paint_heatmap}
\alias{paint_hillshade}
\title{Add paint rules to a layer style.}
\usage{
paint_background(color = "#000000", opacity = 1, pattern = NULL)

paint_fill(
  antialias = TRUE,
  color = "#000000",
  opacity = 1,
  outline_color = NULL,
  pattern = NULL,
  translate = c(0, 0),
  translate_anchor = c("map", "viewport")
)

paint_line(
  blur = 0,
  color = "#000000",
  dasharray = NULL,
  gap_width = 0,
  gradient = NULL,
  offset = 0,
  opacity = 1,
  pattern = NULL,
  translate = c(0, 0),
  translate_anchor = c("map", "viewport"),
  width = 1
)

paint_symbol(
  icon_color = "#000000",
  icon_halo_blur = 0,
  icon_halo_color = "#000000",
  icon_halo_width = 0,
  icon_opacity = 1,
  icon_translate = c(0, 0),
  icon_translate_anchor = c("map", "viewport"),
  text_color = "#000000",
  text_halo_blur = 0,
  text_halo_color = "#000000",
  text_halo_width = 0,
  text_opacity = 1,
  text_translate = c(0, 0),
  text_translate_anchor = c("map", "viewport")
)

paint_raster(
  brightness_max = 1,
  brightness_min = 0,
  contrast = 0,
  fade_duration = 300,
  hue_rotate = 0,
  opacity = 1,
  resampling = c("linear", "nearest"),
  saturation = 0
)

paint_circle(
  blur = 0,
  color = "#000000",
  opacity = 1,
  pitch_alignment = c("map", "viewport"),
  pitch_scale = c("map", "viewport"),
  radius = 5,
  stroke_color = "#000000",
  stroke_opacity = 1,
  stroke_width = 0,
  translate = c(0, 0),
  translate_anchor = c("map", "viewport")
)

paint_fill_extrusion(
  base = 0,
  color = "#000000",
  height = 0,
  opacity = 1,
  pattern = NULL,
  translate = c(0, 0),
  translate_anchor = c("map", "viewport"),
  vertical_gradient = TRUE
)

paint_heatmap(color, intensity = 1, opacity = 1, radius = 30, weight = 1)

paint_hillshade(
  accent_color = "#000000",
  exaggeration = 0.5,
  highlight_color = "#FFFFFF",
  illumination_anchor = c("map", "viewport"),
  illumination_direction = 335,
  shadow_color = "#000000"
)
}
\description{
Add paint rules to a layer style.
}
\details{
For specific parameter descriptions, see the
\href{https://maplibre.org/maplibre-gl-js-docs/style-spec/layers/}{MapLibre GL style specification}.
All paint defaults are the same defaults from the MapLibre GL style specification.
\subsection{Expressions}{

\href{https://maplibre.org/maplibre-gl-js-docs/style-spec/expressions/}{Expressions}
are supported via nested lists, in the same manner that Javascript arrays are used.
}
}
