% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-sf.R
\name{st_geometry}
\alias{st_geometry}
\alias{st_geometry.sf}
\alias{st_geometry.sfc}
\title{Get, set, replace or rename geometry from an sf object}
\usage{
st_geometry(obj, ...)

\method{st_geometry}{sf}(obj, ...)

\method{st_geometry}{sfc}(obj, ...)
}
\arguments{
\item{obj}{object of class \code{sf} or \code{sfc}}

\item{...}{ignored}
}
\value{
st_geometry returns an object of class \link{sfc}, a list-column with geometries
}
\description{
Get, set, replace or rename geometry from an sf object
}
\keyword{internal}
