% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{carto}
\alias{carto}
\alias{stamen}
\alias{openstreetmap}
\title{Basemap Styles}
\usage{
carto(theme = c("dark-matter", "voyager", "positron"))

stamen(theme = c("watercolor", "terrain", "toner"))

openstreetmap()
}
\arguments{
\item{theme}{Theme name for \strong{CARTO} and \strong{Stamen} basemaps}
}
\value{
URL to a TileJSON style file or parameterized raster tiles URL
}
\description{
Basemap Styles
}
