% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_source.R
\name{map_source}
\alias{map_source}
\alias{map_source.default}
\alias{map_source.character}
\alias{map_source.list}
\alias{map_source.sf}
\alias{map_source.sfc}
\alias{map_source.sfg}
\title{Create a map source}
\usage{
map_source(map, x, ..., id = NULL)

\method{map_source}{default}(map, x, ..., id = NULL)

\method{map_source}{character}(map, x, ..., type = source_types(), id = NULL)

\method{map_source}{list}(map, x, ..., type = source_types(), id = NULL)

\method{map_source}{sf}(map, x, ..., id = NULL)

\method{map_source}{sfc}(map, x, ..., id = NULL)

\method{map_source}{sfg}(map, x, ..., id = NULL)
}
\description{
Create a map source
}
